requirements.txt
selenium==3.141.0
webdriver-manager==2.3.0
pyinstaller
pyyaml